nfc3: extract to header
